refactor to support pipelining
this refactor modifies SGEMMCalculator to accept textures for
reading and writing to. It also offloads extraction of data
(into an array) to the caller and allows the floating
point encode step to ommited.
