+++ lib/sgemmcalculator.js
@@ -1,4 +1,4 @@
-
+var WebGL = require('./webgl');
 
 /* A calculator object for the Float texture based GEMM
 
@@ -10,18 +10,25 @@
 
 
 	webgl - a weblas.WebGL object
+	standalone mode compiles in the float encoder (this is required for mobile,
+	circa 2015) but can't be used as part of a pipeline.
 
 	* uploads and downloads data
 	* executes calculation
  */
-function SGEMMCalculator(webgl){
-	this.webgl = webgl;
+function SGEMMCalculator(webgl, standalone){
+	this.webgl = webgl,
+	this.standalone = standalone || true; // default to standalone mode
 
 
 	// create the webgl shader program for this calculation
 	// based on the specific fragment shader for this calculation
 	// and the generic pass through shader
-	this.program = this.webgl.createProgram(SGEMMCalculator.FRAGMENT_SHADER);
+	if(this.standalone){
+		this.program = this.webgl.createProgram(SGEMMCalculator.STANDALONE_FRAGMENT_SHADER);
+	} else {
+		this.program = this.webgl.createProgram(SGEMMCalculator.PIPELINE_FRAGMENT_SHADER);
+	}
 
 }
 
@@ -35,25 +42,13 @@ SGEMMCalculator.TEXTURE_UNIFORM_NAME_1 = "B_t";
 SGEMMCalculator.SHARED_LENGTH_UNIFORM_NAME = "K";
 SGEMMCalculator.ALPHA_UNIFORM_NAME = "alpha";
 
-/* The GLSL fragment shader that carries out the calculation.
-
-   the `outTex` variable is named and defined in the pass-through vertex shader
- */
-SGEMMCalculator.FRAGMENT_SHADER = "                                  \n\
-// fragment shader that calculates the matrix product and renders each   \n\
-// element to the bytes representing a 32-bit IEEE754 floating point in  \n\
-// the output RGBA canvas.                                               \n\
-// readPixel is used to read the bytes.                                  \n\
-																		 \n\
-precision highp float;                                                   \n\
-																		 \n\
-varying vec2      outTex;	// texture coords of row/column to calculate \n\
-uniform sampler2D A;		// texture with data from padded A           \n\
-uniform sampler2D B_t;		// texture with data from padded transpose of B \n\
-uniform int       K;		// number of elements in shared dimension    \n\
-uniform float alpha; 		// coefficient to multiplication             \n\
-float delta = 1./float(K);	// space between items in shared dimension   \n\
-																		 \n\
+SGEMMCalculator.DOT_FUNCTION = "\n\
+ uniform sampler2D A;		// texture with data from padded A           \n\
+ uniform sampler2D B_t;		// texture with data from padded transpose of B \n\
+ uniform int       K;		// number of elements in shared dimension    \n\
+ uniform float alpha; 		// coefficient to multiplication             \n\
+ float delta = 1./float(K);	// space between items in shared dimension   \n\
+ \n\
 // sum of products between elements in row i (from A) x col j (from B)   \n\
 //																		 \n\
 // Calculate the dot product between the row (from A) and column (from B)\n\
@@ -66,41 +61,37 @@ float delta = 1./float(K);	// space between items in shared dimension   \n\
 float dot_rowcol(float i, float j) {                                     \n\
 	float sum = 0.;			// sum for this row/column pair              \n\
 	float k = 0.5 * delta;	// position for shared dimension on source textures \n\
-																		 \n\
+			 															\n\
 	for (int l=0 ; l<4096 ; ++l) {                                       \n\
 		if(l >= K / 4) break;    // stop when we finish the row/column   \n\
-																		 \n\
+				 															\n\
 		// retrieve next four elements from each texture                 \n\
 		vec4 a_ik = texture2D(  A, vec2(k, i));                          \n\
 		vec4 b_kj = texture2D(B_t, vec2(k, j));                          \n\
-																		 \n\
+				 															\n\
 		// use `dot` to process four elements at a time                  \n\
 		sum += alpha * dot(a_ik, b_kj);                                  \n\
 		k += (delta * 4.0);      // (k * 4 + 0.5)*delta                  \n\
 	}                                                                    \n\
 	return sum;                                                          \n\
-}                                                                        \n\
+}";
+
+/* The GLSL fragment shader that carries out the calculation.
+
+   the `outTex` variable is named and defined in the pass-through vertex shader
+ */
+SGEMMCalculator.STANDALONE_FRAGMENT_SHADER = "                                  \n\
+// fragment shader that calculates the matrix product and renders each   \n\
+// element to the bytes representing a 32-bit IEEE754 floating point in  \n\
+// the output RGBA canvas.                                               \n\
+// readPixel is used to read the bytes.                                  \n\
 																		 \n\
-// Render float to bytes according to IEEE 754 Floating Point            \n\
-vec4 encodeFloat(float val) {                                            \n\
-																		 \n\
-	// TODO: correctly handle denormal numbers                           \n\
-	// http://www.2ality.com/2012/04/number-encoding.html                \n\
-	float a = abs(val);                           // encode absolute value + sign \n\
-	float exp = floor(log2(a));                 // number of powers of 2 \n\
-	float mant = pow(2.,log2(a)-exp) * pow(2.,23.);  // multiply to fill 24 bits (implied leading 1) \n\
-	float mant1 = floor(mant / 256. / 256.);    // first 8 bits of mantissa \n\
-	float mant2 = mod(floor(mant / 256.),256.); // second 8 bits         \n\
-	float mant3 = mod(mant,256.);               // third 8 bits          \n\
-																		 \n\
-	highp float sign = 128.-128.*(a/val);			// sign bit is 256 or 0  \n\
-	highp float e = (sign+exp+127.)/510.;		// exponent and sign     \n\
-	highp float m1 = (mant1-(128.*(1.-mod(exp+127.,2.))))/255.; // handle leading bit \n\
-	highp float m2 = (mant2)/255.;				// middle part           \n\
-	highp float m3 = (mant3+.5)/255.;			// scale to 0 - 255      \n\
-																		 \n\
-	return vec4(m3,m2,m1,e);                                             \n\
-}                                                                        \n\
+precision highp float;                                                   \n\
+varying vec2      outTex;	// texture coords of row/column to calculate \n\
+	                                                              		\n" +
+SGEMMCalculator.DOT_FUNCTION +
+WebGL.ENCODE_FUNCTION +
+"																		 \n\
 																		 \n\
 void main(void) {                                                        \n\
 																		 \n\
@@ -123,14 +114,15 @@ void main(void) {                                                        \n\
 }                                                                        \n\
 ";
 
+
 /* Calculate the GEMM, with the given data.
 
 	M - number of rows in A
 	N - number of columns in B
-	K - number of elements in shared dimension
+	K - number of elements in shared dimension (including padding)
 	alpha - scalar for A
-	A - left hand matrix
-	B - right hand matrix
+	A - left hand matrix (as padded texture)
+	B - transpose of right hand matrix (as padded texture)
 	beta - scalar for C
 	C - additive matrix
 
@@ -149,120 +141,33 @@ void main(void) {                                                        \n\
  */
 SGEMMCalculator.prototype.calculate = function(M, N, K, alpha, A, B, beta, C){
 
-	var gl = this.webgl.context,
-		destTexture,
-		frameBuffer,
-		rawbuffer;
+	var gl = this.webgl.context;
 
 	/*
 	var h1 = M, w1 = K,
 		h2 = K, w2 = N;
 	*/
 
-
+	// set this calculator program as the active program
 	this.webgl.selectProgram(this.program);
 
-	// pack each matrix into a single RGBA texel array, with the second transposed
-	var texels0 = SGEMMCalculator.packData(M, K, A, false);
-		texels1 = SGEMMCalculator.packData(K, N, B, true);
-
-
-	var mod = (K % 4),
-		pad = mod == 0 ? 0 : 4 - mod;
-
-	// create and bind our input texture using matrix data
-	this.bindInputTexture(M, K + pad, texels0, gl.TEXTURE0, SGEMMCalculator.TEXTURE_UNIFORM_NAME_0);
-	this.bindInputTexture(N, K + pad, texels1, gl.TEXTURE1, SGEMMCalculator.TEXTURE_UNIFORM_NAME_1);
-
+	//  bind our input textures containing matrix data
+	this.bindInputTexture(A, gl.TEXTURE0, SGEMMCalculator.TEXTURE_UNIFORM_NAME_0);
+	this.bindInputTexture(B, gl.TEXTURE1, SGEMMCalculator.TEXTURE_UNIFORM_NAME_1);
 
 	// set the data specific variables in our shader program
-	this.bindUniforms(K + pad, alpha);
+	this.bindUniforms(K, alpha);
 
-	// create our destination texture
-	destTexture = this.webgl.createDestinationTexture(M, N);
-	frameBuffer = this.webgl.bindDestinationTexture(M, N, destTexture);
-
-
-	if( gl.checkFramebufferStatus(gl.FRAMEBUFFER) != gl.FRAMEBUFFER_COMPLETE)
-		throw new Error("Bound framebuffer is not complete.");
+	// bind our destination texture
+	this.webgl.bindDestinationTexture(M, N, C);
 
 	// initiate calculation
 	gl.drawElements(gl.TRIANGLES, /*num items*/6, gl.UNSIGNED_SHORT, 0);
 
-	// create destination buffer
-	rawbuffer = new ArrayBuffer(M*N*Float32Array.BYTES_PER_ELEMENT);
-
-	// read the result into our buffer, as bytes
-	prod = new Uint8Array(rawbuffer);
-	gl.readPixels(0, 0, N, M, gl.RGBA, gl.UNSIGNED_BYTE, prod);
-
-	// create and return a view over result bytes as a float array
-	return new Float32Array(rawbuffer); // M x N
+	// result can now be read with gl.readResult, or more operations can be
+	// performed on destination texture (in pipeline mode)
 };
 
-/* Pack the given matrix data into texel layout for use by texture shader.
-
-   This layout places consecutive elements of the input data into separate
-   channels, padding to a multiple of four (with zeros) where necessary to fill
-   out the final RGBA texel in a column.
-
-   r - row count
-   c - column count
-   data - TypedArray containing matrix data
-   transpose - whether or not to transpose the data when packing
- */
-SGEMMCalculator.packData = function(r, c, data, transpose) {
-
-	var CHANNELS_PER_TEXEL = 4; // RGBA: four channels, one per color
-
-	var k = !transpose ? c : r;
-
-	var mod = (k % CHANNELS_PER_TEXEL),
-		pad = mod == 0 ? 0 : CHANNELS_PER_TEXEL - mod;
-
-	if (mod === 0 && !transpose) {
-		// special case if column count is a multiple of number of channels
-		return data;
-	}
-
-
-	// dimensions
-	var i, j, p;
-
-	var texelcount = !transpose ? r*(c + pad) : c*(r + pad);
-
-	// create Float32Array to hold padded texel data
-	var texels = new Float32Array(texelcount);
-
-	for(i = 0; i < r; i++){
-
-		if(!transpose){
-			// copy actual data
-			for(j = 0; j < c; j++){
-				texels[i * (c + pad) + j] = data[i * c + j];
-			}
-
-			// pad last texel in this row with zeros
-			for(p = 0; p < pad; p++){
-				texels[i * (c + pad) + j + p] = 0.0;
-			}
-		} else {
-			// copy actual data, transposed
-			for(j = 0; j < c; j++){
-				texels[j * (r + pad) + i] = data[i * c + j];
-			}
-
-			// pad last texel in this row with zeros
-			for(p = 0; p < pad; p++){
-				texels[j * (r + pad) + i + p] = 0.0;
-			}
-
-		}
-
-	}
-
-	return texels;
-};
 
 /* Create a texture from the given texel data and bind it to our shader program.
 
@@ -274,30 +179,16 @@ SGEMMCalculator.packData = function(r, c, data, transpose) {
 
 	must compile program (with createProgram) first
 */
-SGEMMCalculator.prototype.bindInputTexture = function(h, w, texels, textureUnit, name){
+SGEMMCalculator.prototype.bindInputTexture = function(texture, textureUnit, name){
 	var gl = this.webgl.context,
 		program = this.program;
 
-
-
-	// create the texture from our floats
-	var texture = gl.createTexture();
 	gl.activeTexture(textureUnit); // gl.TEXTURE0, gl.TEXTURE1, etc
 	gl.bindTexture(	  gl.TEXTURE_2D, texture);
-	gl.texImage2D(	  gl.TEXTURE_2D, /*level*/0, gl.RGBA, w / 4, h, 0,
-					  gl.RGBA, gl.FLOAT, texels);
-
-	// clamp to edge to support non-power of two textures
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-	// don't interpolate when getting data from texture
-	gl.texParameteri( gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
-	gl.texParameteri( gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
 
 	var sampler = gl.getUniformLocation(program, name);
 	gl.uniform1i(sampler, textureUnit - gl.TEXTURE0);
 
-	return texture;
 };
 
 /* Set up inputs for the texture shader
@@ -308,11 +199,12 @@ SGEMMCalculator.prototype.bindUniforms = function(K, alpha) {
 	var gl = this.webgl.context;
 
 	// get var locations
-	var K_gl	= gl.getUniformLocation(this.program, SGEMMCalculator.SHARED_LENGTH_UNIFORM_NAME),
+	var K_gl	 = gl.getUniformLocation(this.program, SGEMMCalculator.SHARED_LENGTH_UNIFORM_NAME),
 		alpha_gl = gl.getUniformLocation(this.program, SGEMMCalculator.ALPHA_UNIFORM_NAME);
 
 	// bind length of shared dimension
 	gl.uniform1i(K_gl, K);
+	// bind alpha
 	gl.uniform1f(alpha_gl, alpha);
 
 };
