+++ lib/webgl.js
@@ -85,6 +85,28 @@ void main(void)                                                              \n\
 }                                                                            \n\
 ";
 
+WebGL.ENCODE_FUNCTION = "\n\
+// Render float to bytes according to IEEE 754 Floating Point            \n\
+vec4 encodeFloat(float val) {                                            \n\
+																	 \n\
+// TODO: correctly handle denormal numbers                           \n\
+// http://www.2ality.com/2012/04/number-encoding.html                \n\
+float a = abs(val);                           // encode absolute value + sign \n\
+float exp = floor(log2(a));                 // number of powers of 2 \n\
+float mant = pow(2.,log2(a)-exp) * pow(2.,23.);  // multiply to fill 24 bits (implied leading 1) \n\
+float mant1 = floor(mant / 256. / 256.);    // first 8 bits of mantissa \n\
+float mant2 = mod(floor(mant / 256.),256.); // second 8 bits         \n\
+float mant3 = mod(mant,256.);               // third 8 bits          \n\
+																	 \n\
+highp float sign = 128.-128.*(a/val);			// sign bit is 256 or 0  \n\
+highp float e = (sign+exp+127.)/510.;		// exponent and sign     \n\
+highp float m1 = (mant1-(128.*(1.-mod(exp+127.,2.))))/255.; // handle leading bit \n\
+highp float m2 = (mant2)/255.;				// middle part           \n\
+highp float m3 = (mant3+.5)/255.;			// scale to 0 - 255      \n\
+																	 \n\
+return vec4(m3,m2,m1,e);                                             \n\
+}";
+
 /*  Create a shader program based on a pass through vertex shader and
 	the supplied operation specific fragment shader.
 
@@ -168,8 +190,37 @@ WebGL.prototype.bindVertices = function(program) {
 	gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(vertexIndices), gl.STATIC_DRAW);
 };
 
-/* Create and bind a texture to store the result.
-   Requires canvas height and width be set to size of output matrix.
+/* create RGBA texture of width w/4 from given texels
+   bind to texture unit and set as uniform in texture shader program to use
+ */
+WebGL.prototype.createInputTexture = function(h, w, texels){
+
+	var gl = this.context;
+
+	// create the texture from our floats
+	var texture = gl.createTexture();
+	gl.bindTexture(	  gl.TEXTURE_2D, texture);
+	gl.texImage2D(	  gl.TEXTURE_2D, /*level*/0, gl.RGBA, w / 4, h, 0,
+					  gl.RGBA, gl.FLOAT, texels);
+
+	// clamp to edge to support non-power of two textures
+	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
+	gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
+	// don't interpolate when getting data from texture
+	gl.texParameteri( gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
+	gl.texParameteri( gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
+
+	// we're done with setup, so unbind current texture
+	gl.bindTexture(gl.TEXTURE_2D, null);
+
+	return texture;
+};
+
+/* Create a texture to store the result.
+   Can be passed to bindDestinationTexture.
+   
+   Returns an unsigned byte RGBA texture (other formats are not yet supported
+	on most platforms, see WEBGL_color_buffer_float extension)
  */
 WebGL.prototype.createDestinationTexture = function(h, w) {
 	var gl = this.context;
@@ -196,7 +247,7 @@ WebGL.prototype.createDestinationTexture = function(h, w) {
 	N - number of columns in output
 	dstTex - texture for holding the output
  */
-WebGL.prototype.bindDestinationTexture = function(M, N, dstTex) {
+WebGL.prototype.bindDestinationTexture = function(M, N, texture) {
 	var gl = this.context;
 
 	// set canvas and viewport size
@@ -204,21 +255,30 @@ WebGL.prototype.bindDestinationTexture = function(M, N, dstTex) {
 	this.canvas.width = N;
 	gl.viewport(0, 0, N, M);
 
-	// create and bind renderbuffer
-	this.renderbuffer = this.renderbuffer || gl.createRenderbuffer();
-
-	gl.bindRenderbuffer(gl.RENDERBUFFER, null);
-	gl.bindRenderbuffer(gl.RENDERBUFFER, this.renderbuffer);
-
-	gl.renderbufferStorage(gl.RENDERBUFFER,gl.DEPTH_COMPONENT16, N, M);
-
 	// create and bind framebuffer
 	this.framebuffer = this.framebuffer || gl.createFramebuffer();
 
 	gl.bindFramebuffer(gl.FRAMEBUFFER, this.framebuffer);
 
-	gl.framebufferTexture2D(gl.FRAMEBUFFER,gl.COLOR_ATTACHMENT0,gl.TEXTURE_2D,dstTex,/*level*/0);
-	gl.framebufferRenderbuffer(gl.FRAMEBUFFER,gl.DEPTH_ATTACHMENT,gl.RENDERBUFFER,this.renderbuffer);
+	gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, texture, /*level*/0);
+
+	if( gl.checkFramebufferStatus(gl.FRAMEBUFFER) != gl.FRAMEBUFFER_COMPLETE)
+		throw new Error("Bound framebuffer is not complete.");
 
 	return this.framebuffer;
 };
+
+/* Read data out as unsigned bytes */
+WebGL.prototype.readData = function(M, N){
+	var gl = this.context;
+
+	// create destination buffer
+	rawbuffer = new ArrayBuffer(M*N*Float32Array.BYTES_PER_ELEMENT);
+
+	// read the result into our buffer, as bytes
+	prod = new Uint8Array(rawbuffer);
+	gl.readPixels(0, 0, N, M, gl.RGBA, gl.UNSIGNED_BYTE, prod);
+
+	// create and return a view over result bytes as a float array
+	return new Float32Array(rawbuffer); // M x N
+}
